refacor: divide components and css modify
