+++ client/src/components/community/FilterTab.tsx
@@ -7,6 +7,7 @@ const FilterDiv = styled.div<{ selected: boolean }>`
   padding-bottom: 0.5rem;
   border-bottom: 0.375rem solid
     ${(props) => (props.selected ? cssToken.COLOR['point-900'] : 'transparent')};
+  cursor: pointer;
 `;
 const FilterTab = ({
   selectTab,
